requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.2
fonttools==4.54.1
GitPython==3.1.43
gitdb==4.0.11
idna==3.10
imageio==2.35.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.9.1
networkx==3.3
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pre-commit==4.0.1
protobuf==5.28.2
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
ruff==0.6.9
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
smmap==5.0.1
streamlit==1.39.0
tenacity==9.0.0
tifffile==2024.9.20
toml==0.10.2
torch==2.0.0
tornado==6.4.1
transformers==4.37.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==5.0.3
